move stack-to-shared pass from Flang to the OpenMP dialect
