+++ mlir/lib/Dialect/OpenMP/IR/CMakeLists.txt
@@ -1,24 +1,18 @@
-add_mlir_dialect_library(MLIROpenMPTransforms
-  OpenMPOffloadPrivatizationPrepare.cpp
-  StackToShared.cpp
+add_mlir_dialect_library(MLIROpenMPDialect
+  OpenMPDialect.cpp
 
   ADDITIONAL_HEADER_DIRS
   ${MLIR_MAIN_INCLUDE_DIR}/mlir/Dialect/OpenMP
 
   DEPENDS
   omp_gen
-  MLIROpenMPPassIncGen
   MLIROpenMPOpsIncGen
   MLIROpenMPOpsInterfacesIncGen
   MLIROpenMPTypeInterfacesIncGen
 
   LINK_LIBS PUBLIC
   MLIRIR
-  MLIRFuncDialect
   MLIRLLVMDialect
+  MLIRFuncDialect
   MLIROpenACCMPCommon
-  MLIROpenMPDialect
-  MLIRPass
-  MLIRSupport
-  MLIRTransforms
   )
